+++ Assets/Bookmark4Unity/Editor/SceneObjectBookmarkGroup.cs
@@ -1,5 +1,6 @@
 namespace Bookmark4Unity.Editor
 {
+    using System;
     using System.Collections.Generic;
     using Bookmark4Unity.Guid;
     using UnityEditor;
@@ -10,17 +11,11 @@ namespace Bookmark4Unity.Editor
     {
         public VisualElement Element { get; private set; }
         public Foldout Root { get; private set; }
-        public List<GuidReference> Data { get; private set; }
-        public ListView DataListView { get; private set; }
-        public const int ItemHeight = 15;
-        public bool IsEmpty => Data.Count < 1;
-        public Dictionary<int, EventCallback<ClickEvent>> pingActions = new();
-        public Dictionary<int, EventCallback<ClickEvent>> focusActions = new();
-        public Dictionary<int, EventCallback<ClickEvent>> delActions = new();
-        public Dictionary<int, EventCallback<PointerLeaveEvent>> dragActions = new();
+        public SceneObjectListView SceneObjListView { get; private set; }
+        public SceneObjectCollectionListView SceneObjCollectionListView { get; private set; }
 
 
-        public SceneObjectBookmarkGroup(string groupName, Color borderColor, List<GuidReference> data, VisualTreeAsset groupAsset, VisualTreeAsset btnAsset)
+        public SceneObjectBookmarkGroup(string groupName, Color borderColor, List<GuidReference> objReferenceData, List<SceneObjectReferenceCollection> objCollectionData, VisualTreeAsset groupAsset, VisualTreeAsset btnAsset)
         {
             Element = groupAsset.Instantiate();
             Root = Element.Q<Foldout>("Root");
@@ -29,148 +24,60 @@ namespace Bookmark4Unity.Editor
             Root.style.borderLeftColor = borderColor;
             Root.style.borderRightColor = borderColor;
             Root.text = groupName;
-            Data = data;
-
-            DataListView = new(Data, ItemHeight, () =>
-            {
-                return btnAsset.Instantiate();
-            },
-            (item, i) =>
-            {
-                var icon = item.Q<Button>("Icon");
-                var btn = item.Q<Button>("Btn");
-                var focus = item.Q<Button>("Focus");
-                var del = item.Q<Button>("Del");
-                var data = Data[i];
-                var index = i; // save value for lambda functions
-                icon.style.backgroundImage = Background.FromTexture2D(
-                    data.gameObject is null ?
-                    EditorGUIUtility.IconContent("console.warnicon").image as Texture2D :
-                    PrefabUtility.GetIconForGameObject(data.gameObject));
-                focus.style.backgroundImage = Background.FromTexture2D(SceneViewBookmarkManager.SceneViewBookmarkIcon);
-                del.style.backgroundImage = Background.FromTexture2D(SceneViewBookmarkManager.SceneViewEmptyIcon);
-                btn.text = data.CachedName;
-
-                // ping
-                if (pingActions.ContainsKey(i)) btn.UnregisterCallback<ClickEvent>(pingActions[i]);
-                pingActions[i] = evt => Ping(index);
-                btn.RegisterCallback<ClickEvent>(pingActions[i]);
-                btn.tooltip = $"Select \"{data.CachedName}\"";
-
-                // focus
-                if (focusActions.ContainsKey(i)) focus.UnregisterCallback<ClickEvent>(focusActions[i]);
-                focusActions[i] = evt => Focus(index);
-                focus.RegisterCallback<ClickEvent>(focusActions[i]);
-                focus.tooltip = $"Focus on \"{data.CachedName}\"";
-
-                // del
-                if (delActions.ContainsKey(i)) del.UnregisterCallback<ClickEvent>(delActions[i]);
-                delActions[i] = evt => Remove(index);
-                del.RegisterCallback<ClickEvent>(delActions[i]);
-                del.tooltip = $"Unpin \"{data.CachedName}\"";
-
-                // drag
-                if (dragActions.ContainsKey(i)) btn.UnregisterCallback<PointerLeaveEvent>(dragActions[i]);
-                dragActions[i] = evt => OnDrag(index);
-                btn.RegisterCallback<PointerLeaveEvent>(dragActions[i]);
-            })
-            {
-                reorderable = true,
-                showBorder = false
-            };
-
-            DataListView.style.flexGrow = 1f; // Fills the window
-            Root.Add(DataListView);
+            SceneObjListView = new SceneObjectListView(objReferenceData, btnAsset);
+            SceneObjListView.ChangeEvent += OnListViewChange;
+            SceneObjCollectionListView = new SceneObjectCollectionListView(objCollectionData, btnAsset);
+            SceneObjCollectionListView.ChangeEvent += OnListViewChange;
+            Root.Add(SceneObjCollectionListView.DataListView);
+            Root.Add(SceneObjListView.DataListView);
         }
 
-        public bool Add(GuidReference data)
+        public bool AddSceneObject(GuidReference data)
         {
-            if (Data.Contains(data)) return false;
-            Data.Add(data);
-            DataListView.Rebuild();
-            Element.RemoveFromClassList(Bookmark4UnityWindow.HiddenContentClassName);
-            return true;
+            return SceneObjListView.Add(data);
         }
 
-        public void Ping(int index)
+        public bool AddSceneObjectCollection(SceneObjectReferenceCollection data)
         {
-            if (Data[index].gameObject != null)
-            {
-                Selection.activeGameObject = Data[index].gameObject;
-            }
-            else
-            {
-                if (EditorUtility.DisplayDialog(Data[index].CachedName, "Selected game object dows not exist on current scene, remove it from list?", "Yes", "No"))
-                {
-                    Remove(index);
-                }
-            }
-        }
-
-        public void Focus(int index)
-        {
-            if (Data[index].gameObject != null)
-            {
-                Selection.activeGameObject = Data[index].gameObject;
-                SceneView.lastActiveSceneView.FrameSelected();
-            }
-            else
-            {
-                if (EditorUtility.DisplayDialog(Data[index].CachedName, "Selected game object dows not exist on current scene, remove it from list?", "Yes", "No"))
-                {
-                    Remove(index);
-                }
-            }
-        }
-
-        public void Remove(int index)
-        {
-            if (Data[index].gameObject is not null)
-            {
-                UnityEngine.Object.DestroyImmediate(Data[index].gameObject.GetComponent<GuidComponent>());
-            }
-
-            Data.RemoveAt(index);
-            DataListView.Rebuild();
-            if (IsEmpty) Element.AddToClassList(Bookmark4UnityWindow.HiddenContentClassName);
-            Bookmark4UnityWindow.UpdateSavedData();
+            return SceneObjCollectionListView.Add(data);
         }
 
         public void RemoveAll()
         {
-            for (int i = 0; i < Data.Count; i++)
-            {
-                if (Data[i].gameObject is not null)
-                {
-                    UnityEngine.Object.DestroyImmediate(Data[i].gameObject.GetComponent<GuidComponent>());
-                }
-            }
-
-            Data.Clear();
-            DataListView.Rebuild();
+            SceneObjListView.RemoveAll();
+            SceneObjCollectionListView.RemoveAll();
             Element.AddToClassList(Bookmark4UnityWindow.HiddenContentClassName);
             Bookmark4UnityWindow.UpdateSavedData();
         }
 
-        private void OnDrag(int index)
-        {
-            if (Event.current.type != EventType.MouseDrag || Data[index].gameObject is null) return;
-            DragAndDrop.PrepareStartDrag();
-            DragAndDrop.objectReferences = new UnityEngine.Object[] { Data[index].gameObject };
-            DragAndDrop.StartDrag(Data[index].CachedName);
-            Event.current.Use();
-        }
-
         public void SortDesc()
         {
-            Data.Sort((a, b) => a.CachedName.CompareTo(b.CachedName));
-            DataListView.RefreshItems();
+            SceneObjListView.SortDesc();
+            SceneObjCollectionListView.SortDesc();
         }
 
         public void SortAsc()
         {
-            Data.Sort((a, b) => b.CachedName.CompareTo(a.CachedName));
-            DataListView.RefreshItems();
+            SceneObjListView.SortAsc();
+            SceneObjCollectionListView.SortAsc();
+        }
+
+        public void Refresh()
+        {
+            SceneObjListView.DataListView.RefreshItems();
+            SceneObjCollectionListView.DataListView.RefreshItems();
+        }
+
+        private void OnListViewChange()
+        {
+            if (SceneObjListView.IsEmpty && SceneObjCollectionListView.IsEmpty)
+            {
+                Element.AddToClassList(Bookmark4UnityWindow.HiddenContentClassName);
+            }
+            else
+            {
+                Element.RemoveFromClassList(Bookmark4UnityWindow.HiddenContentClassName);
+            }
         }
     }
 }