+++ Assets/Bookmark4Unity/Editor/Utility/EditorInputDialog.cs
@@ -1,0 +1,109 @@
+namespace Bookmark4Unity.Editor
+{
+    using System;
+    using UnityEditor;
+    using UnityEngine;
+
+    /// <summary>
+    /// Modified Vedran_M's shared code.
+    /// https://forum.unity.com/threads/is-there-a-way-to-input-text-using-a-unity-editor-utility.473743/
+    /// </summary>
+    public class EditorInputDialog : EditorWindow
+    {
+        string description, inputText;
+        string okButton, cancelButton;
+        Action onOKButton;
+
+        bool shouldClose = false;
+
+        void OnGUI()
+        {
+            // Check if Esc/Return have been pressed
+            var e = Event.current;
+            if (e.type == EventType.KeyDown)
+            {
+                switch (e.keyCode)
+                {
+                    // Escape pressed
+                    case KeyCode.Escape:
+                        shouldClose = true;
+                        e.Use();
+                        break;
+
+                    // Enter pressed
+                    case KeyCode.Return:
+                    case KeyCode.KeypadEnter:
+                        onOKButton?.Invoke();
+                        shouldClose = true;
+                        e.Use();
+                        break;
+                }
+            }
+
+            if (shouldClose)
+            {  // Close this dialog
+                Close();
+            }
+
+            // Draw our control
+            var rect = EditorGUILayout.BeginVertical();
+
+            EditorGUILayout.Space(12);
+            EditorGUILayout.LabelField(description);
+
+            EditorGUILayout.Space(8);
+            GUI.SetNextControlName("inText");
+            inputText = EditorGUILayout.TextField("", inputText);
+            GUI.FocusControl("inText");   // Focus text field
+            EditorGUILayout.Space(12);
+
+            // Draw OK / Cancel buttons
+            var r = EditorGUILayout.GetControlRect();
+            r.width /= 2;
+            if (GUI.Button(r, okButton))
+            {
+                onOKButton?.Invoke();
+                shouldClose = true;
+            }
+
+            r.x += r.width;
+            if (GUI.Button(r, cancelButton))
+            {
+                inputText = null;   // Cancel - delete inputText
+                shouldClose = true;
+            }
+
+            EditorGUILayout.Space(8);
+            EditorGUILayout.EndVertical();
+
+            // Force change size of the window
+            if (rect.width != 0 && minSize != rect.size)
+            {
+                minSize = maxSize = rect.size;
+            }
+        }
+
+        /// <summary>
+        /// Returns text player entered, or null if player cancelled the dialog.
+        /// </summary>
+        /// <param name="title"></param>
+        /// <param name="description"></param>
+        /// <param name="inputText"></param>
+        /// <param name="okButton"></param>
+        /// <param name="cancelButton"></param>
+        /// <returns>input string</returns>
+        public static string Show(string title, string description, string inputText, string okButton = "OK", string cancelButton = "Cancel")
+        {
+            string ret = null;
+            var window = CreateInstance<EditorInputDialog>();
+            window.titleContent = new GUIContent(title);
+            window.description = description;
+            window.inputText = inputText;
+            window.okButton = okButton;
+            window.cancelButton = cancelButton;
+            window.onOKButton += () => ret = window.inputText;
+            window.ShowModalUtility();
+            return ret;
+        }
+    }
+}